Fixed review comments on PR 391, restructured to align with exchange new code structure
