+++ authbridge/authlib/plugins/registry.go
@@ -116,5 +116,3 @@ func Build(entries []config.PluginEntry, opts ...pipeline.Option) (*pipeline.Pip
 	}
 	return pipeline.New(ps, opts...)
 }
-
-// Made with Bob
